Change properties

+++ BancoChu/BancoChu.Application/BankAccount/Commands/CreateBankAccountCommand.cs
@@ -5,4 +5,4 @@ namespace BancoChu.Application.BankAccount.Commands;
 public sealed record CreateBankAccountCommand(
     int BranchCode,
     int CurrentAccountNumber,
-    int AccountBalance) : ICommand<Guid>;
+    decimal AccountBalance) : ICommand<Guid>;
